+++ src/utils.rs
@@ -64,7 +64,7 @@ fn verify_internal(a: &[u8], b: &[u8]) -> u8 {
 /// wipe(&mut secret);
 /// ```
 pub fn wipe(secret: &mut [u8]) {
-    unsafe { ffi::crypto_wipe(secret.as_mut_ptr() as *mut c_void, secret.len() as u64) }
+    unsafe { ffi::crypto_wipe(secret.as_mut_ptr() as *mut c_void, secret.len()) }
 }
 
 #[cfg(test)]
